% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.logis_fe.R
\name{coef.logis_fe}
\alias{coef.logis_fe}
\title{Return the model coefficients of a \code{logis_fe} object}
\usage{
\method{coef}{logis_fe}(fit, ...)
}
\arguments{
\item{fit}{an object as output of \code{logis_fe} function.}

\item{...}{}
}
\description{
Return the model coefficients of a \code{logis_fe} object
}
\examples{
data(data_FE)
data.prep <- fe_data_prep(data_FE$Y, data_FE$Z, data_FE$ID, message = FALSE)
fit_fe <- logis_fe(data.prep)
coef(fit_fe)
}
