% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{data_FE}
\alias{data_FE}
\title{Example dataset}
\format{
A list containing the following elements:
\describe{
  \item{Y}{a vector represents the binary outcome}
  \item{ID}{a vector represents the facility indicator (10 facilities in total)}
  \item{Z}{a data frame contains 4 continuous variables}
}
}
\usage{
data(data_FE)
}
\description{
A simulated data set containing response variable, provider information and 4 continuous covariates.
}
\examples{
data(data_FE)
head(data_FE$Y)
head(data_FE$ID)
head(data_FE$Z)

}
\keyword{datasets}
