% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe_data_prep.R
\name{fe_data_prep}
\alias{fe_data_prep}
\title{Data preparation before modelling}
\usage{
fe_data_prep(Y, Z, ID, cutoff = 10, check = TRUE, message = TRUE)
}
\arguments{
\item{Y}{a numerical vector, with values of 0 or 1, indicating the outcome variable.}

\item{Z}{a matrix or data frame containing covariates.}

\item{ID}{a vector representing the provider id. Its elements can be either numeric values or characters.}

\item{cutoff}{an integer as cutoff of provider size with 10 as default. Facilities with observations fewer than the "cutoff" value will be labeled as "include = 0".}

\item{check}{a Boolean indicating whether checking missingness, variation, correlation and VIF of variables in the data. Defaulting to "TRUE".}

\item{message}{a Boolean indicating whether printing out the information of the data preparation process. Defaulting to "TRUE".}

\item{...}{}
}
\value{
\item{data}{a data frame including response, provider ID, and covariates}

\item{char_list}{a list including variable names.}
}
\description{
Data preparation before modelling
}
\details{
Major steps in this stage include (in order):
  \itemize{
  \item checking missingness, variation, correlation and VIF of variables in the data,
  \item provider screening based on number of discharges, 10 as default,
  \item calculating the total number of facilities and hospitals,
  \item reporting proportions of facilities with no readmission within 30 days, and with all readmission within 30 days after discharge,
  \item modifying responses to prevent infinite log-likelihoods during model fitting, and
  \item sorting data by hospital and provider identifiers.
  }


The `fe_data_prep()` function not only returns the original data sorted by provider identifiers,
but also outputs vectors that indicate whether the provider has zero or all readmissions, and
a dataframe combines all the output information is also provided.
}
\examples{
data(data_FE)
data.prep <- fe_data_prep(data_FE$Y, data_FE$Z, data_FE$ID)
head(data.prep$data)
data.prep$char_list

}
\keyword{Data}
\keyword{Effects}
\keyword{Fixed}
\keyword{Preparation,}
\keyword{Provider}
