% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SR_output.R
\name{SR_output}
\alias{SR_output}
\title{Provide direct/indirect standardization ratio or rate}
\usage{
SR_output(
  fit,
  stdz = "indirect",
  measure = c("rate", "ratio"),
  null = "median",
  Rcpp = TRUE,
  threads = 1,
  ...
)
}
\arguments{
\item{fit}{an object as output of \code{logis_fe} function.}

\item{stdz}{a character string specifying the standardization method. Defaulting to "indirect".
\itemize{
\item{"indirect":} using indirect standardized method
\item{"direct":} using direct standardized method
}}

\item{measure}{a boolean indicating whether the output measure is "ratio" or "rate". Both "rate" and "ratio" will be provided by default.
\itemize{
\item{"rate":} output the standardized rate. The "rate" has been restricted to 0\% - 100\%.
\item{"ratio":} output the standardized ratio.
}}

\item{null}{if "stdz = indirect", a character string or real number specifying null hypotheses of fixed provider effects for calculating standardized rate/ratio. Defaulting to "median".}

\item{Rcpp}{a boolean indicating whether to use Rcpp. Defaulting to TRUE.}

\item{threads}{an integer specifying the number of threads to use if "Rcpp = T". Defaulting to 1.}

\item{...}{}
}
\value{
The return values depend on the user's choice of standardization method and measure type

\item{indirect.ratio}{a vector of standardization ratio using indirect method}

\item{direct.ratio}{a vector of standardization ratio using direct method}

\item{indirect.rate}{a vector of standardization rate using indirect method}

\item{direct.rate}{a vector of standardization rate using direct method}
}
\description{
Provide direct/indirect standardization ratio or rate
}
\details{
The \code{"stdz"} and \code{"measure"} arguments must be explicitly provided.
Users are allowed to specify \verb{"stdz = c("indirect", "direct")"} or \verb{"measure = c("rate", "ratio")"} to get multiple measures.
}
\examples{
data(data_FE)
data.prep <- fe_data_prep(data_FE$Y, data_FE$Z, data_FE$ID, message = FALSE)
fit_fe <- logis_fe(data.prep)
SR <- SR_output(fit_fe, stdz = "direct", measure = "rate")
SR$direct.rate

}
\references{
\itemize{
\item He K, Kalbfleisch, J, Li, Y, and et al. (2013) Evaluating hospital readmission rates in dialysis facilities; adjusting for hospital effects.
\emph{Lifetime Data Analysis}, \strong{19}: 490-512.
}
}
