% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_fe.R
\name{test_fe}
\alias{test_fe}
\title{Hypothesis Tests for Identifying Outlying Providers}
\usage{
test_fe(
  fit,
  parm,
  level = 0.95,
  test = "exact.poisbinom",
  null = "median",
  n = 10000
)
}
\arguments{
\item{fit}{an object as output of \code{logis_fe} function}

\item{parm}{specify a subset of which providers are to be given confidence intervals.}

\item{level}{confidence level used for constructing confidence intervals. Defaulting to 0.95}

\item{test}{a character string specifying the type of testing method to be conducted, defaulting to "exact.poisbinom"
* `exact.poisbinom`: two-sided exact test based on Poisson-binomial distribution of \eqn{O_i|Z_i}
* `exact.bootstrap`: two-sided exact test based on bootstrap procedure
* `wald`: wald test
* `score`: score test}

\item{null}{a character string or real number specifying null hypotheses of fixed provider effects}

\item{n}{resample size for bootstrapping, 10000 as default}

\item{...}{}
}
\value{
a dataframe containing:

\item{flag}{a vector of flagging indicator. "1" means better, and "-1" means worse}

\item{p}{p-value}

\item{stat}{z-score}
}
\description{
Hypothesis Tests for Identifying Outlying Providers
}
\details{
By default, the function uses the "exact.poisbinom" method.
The wald test is invalid for extreme providers (i.e. when provider effect goes to infinity).
}
\examples{
data(data_FE)
data.prep <- fe_data_prep(data_FE$Y, data_FE$Z, data_FE$ID)
fit_fe <- logis_fe(data.prep)
test_fe(fit_fe)

}
\references{
\itemize{
\item Wu, W, Yang, Y, Kang, J, He, K. (2022) Improving large-scale estimation and inference for profiling health care providers.
\emph{Statistics in Medicine}, \strong{41(15)}: 2840-2853.

\item He K, Kalbfleisch, J, Li, Y, and et al. (2013) Evaluating hospital readmission rates in dialysis facilities; adjusting for hospital effects.
\emph{Lifetime Data Analysis}, \strong{19}: 490-512.
}
}
