% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe_data_prep.R
\name{fe_data_prep}
\alias{fe_data_prep}
\title{Data preparation before modelling}
\usage{
fe_data_prep(Y, Z, ID, cutoff = 10, check = TRUE, message = TRUE)
}
\arguments{
\item{Y}{a numerical vector, with values of 0 or 1, indicating the outcome variable.}

\item{Z}{a matrix or data frame containing covariates.}

\item{ID}{a vector representing the provider id. Its elements can be either numeric values or characters.}

\item{cutoff}{an integer as cutoff of provider size with 10 as default. Providers with observations fewer than the "cutoff" value will be labeled as "include = 0".}

\item{check}{a Boolean indicating whether checking missingness, variation, correlation and VIF of variables in the data. Defaulting to "TRUE".}

\item{message}{a Boolean indicating whether printing out the information of the data preparation process. Defaulting to "TRUE".}

\item{...}{}
}
\value{
\item{data}{a sorted data frame including response, provider identifiers, covariates, and additional provider information.}

\item{char_list}{a list including variable names.}
}
\description{
Data preparation before modelling
}
\details{
Major steps in this stage include (in order):
\itemize{
\item checking missingness, variation, correlation and VIF of variables in the data,
\item provider screening based on the provider size, 10 as default,
\item reporting proportions of providers with no events (i.e. all "0" outcomes) and with all events (i.e. all "1" outcomes),
\item sorting data by provider identifiers.
}

The \code{fe_data_prep()} function returns data sorted by the provider identifiers,
accompanied by additional provider-related information indicating whether the provider's size is smaller than a specified "cutoff"
and whether the respective provider has experienced either zero or all events.
The reason behind introducing a "cutoff" lies in findings from both simulated and real data studies,
revealing the instability of coefficient estimates for providers with small sizes.
Consequently, we recommend excluding small providers during the model fitting process.
It is important to note that the resultant data frame retains all providers, including small ones,
but utilizes the "included = 0" label to signify the small providers. Subsequently, during the model fitting stage,
the \code{logis_fe()} function disregards records marked with "included = 0".
}
\examples{
data(data_FE)
data.prep <- fe_data_prep(data_FE$Y, data_FE$Z, data_FE$ID)
head(data.prep$data)
data.prep$char_list

}
\keyword{Data}
\keyword{Effects}
\keyword{Fixed}
\keyword{Preparation,}
\keyword{Provider}
