% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logis_fe.R
\name{logis_fe}
\alias{logis_fe}
\title{Main function for fitting fixed effects model}
\usage{
logis_fe(
  data.prep,
  algorithm = "SerBIN",
  max.iter = 10000,
  tol = 1e-05,
  bound = 10,
  backtrack = FALSE,
  Rcpp = TRUE,
  AUC = FALSE,
  message = FALSE
)
}
\arguments{
\item{data.prep}{an object from the \code{fe_data_prep()} function.}

\item{algorithm}{a string specifying the algorithm to be used. Defaulting to "SerBIN".
\itemize{
\item "SerBIN": using the Serial blockwise inversion Newton algorithm to fit the model (See \href{https://onlinelibrary.wiley.com/doi/full/10.1002/sim.9387}{Wu et al. (2022)}).
\item "BAN": using the block ascent Newton algorithm to fit the model (See \href{https://link.springer.com/article/10.1007/s10985-013-9264-6}{He et al. (2013)}).
}}

\item{max.iter}{maximum number of iterations. Defaulting to 10,000.}

\item{tol}{a small positive number specifying stopping criterion of Newton-Raphson algorithm. Defaulting to 1e-5.}

\item{bound}{a positive number to avoid inflation of provider effect. Defaulting to 10.}

\item{backtrack}{a boolean indicating whether backtracking line search is implemented. Defaulting to FALSE.}

\item{Rcpp}{a Boolean indicating whether the Rcpp function is used. Defaulting to TRUE.}

\item{AUC}{a Boolean indicating whether report AUC. Defaulting to FALSE.}

\item{message}{a Boolean indicating whether track the fitting process. Defaulting to TRUE.}

\item{...}{}
}
\value{
An object with S3 class \code{logis_fe}.

\item{beta}{a vector of fixed effects estimates of covariates}

\item{gamma}{a vector of estimates of provider effects}

\item{obs}{a vector of patients-level outcome}

\item{neg2Loglkd}{minus two times log likelihood}

\item{AIC}{Akaike info criterion}

\item{BIC}{Bayesian info criterion}

\item{AUC}{area under the ROC curve}
}
\description{
Main function for fitting fixed effects model
}
\details{
The default algorithm is based on Serial blockwise inversion Newton (SerBIN) proposed by
\href{https://onlinelibrary.wiley.com/doi/full/10.1002/sim.9387}{Wu et al. (2022)},
but users can also choose to use the block ascent Newton (BAN) algorithm proposed by
\href{https://link.springer.com/article/10.1007/s10985-013-9264-6}{He et al. (2013)} to fit the model.
Both methodologies build upon the Newton-Raphson method, yet SerBIN simultaneously updates both the provider effect and covariate coefficient.
This concurrent update necessitates the inversion of the complete information matrix at each iteration.
In contrast, BAN adopts a two-layer updating approach, where the covariate coefficient is sequentially fixed to update the provider effect,
followed by fixing the provider effect to update the covariate coefficient.

We suggest using the default \code{"SerBIN"} option as it typically converges much faster for most datasets.
However, in rare cases where the SerBIN algorithm encounters second-order derivative irreversibility leading to an error,
users can consider using the \code{"BAN"} option as an alternative.

For a deeper understanding, please consult the original article for detailed insights.
}
\examples{
data(data_FE)
data.prep <- fe_data_prep(data_FE$Y, data_FE$Z, data_FE$ID, message = FALSE)
fit_fe <- logis_fe(data.prep)

}
\references{
\itemize{
\item Wu, W, Yang, Y, Kang, J, He, K. (2022) Improving large-scale estimation and inference for profiling health care providers.
\emph{Statistics in Medicine}, \strong{41(15)}: 2840-2853.

\item He K, Kalbfleisch, J, Li, Y, and et al. (2013) Evaluating hospital readmission rates in dialysis providers; adjusting for hospital effects.
\emph{Lifetime Data Analysis}, \strong{19}: 490-512.
}
}
\keyword{(CENR)}
\keyword{Algorithm,}
\keyword{Cost-Efficient}
\keyword{Effects}
\keyword{Fixed}
\keyword{Newton-Raphson}
\keyword{Provider}
